+++ meta/meta_vault.go
@@ -6,13 +6,11 @@ func AdditionalOptionsUsage() string {
 	return `
   -wrap-ttl=""            Indicates that the response should be wrapped in a
                           cubbyhole token with the requested TTL. The response
-                          will live at "cubbyhole/response" in the cubbyhole of
-                          the returned token with a key of "response" and can
-                          be parsed as a normal API Secret. The backend can
-                          also request wrapping; the lesser of the values is
-                          used. This is a numeric string with an optional
-                          suffix "s", "m", or "h"; if no suffix is specified it
-                          will be parsed as seconds. May also be specified via
+                          can be fetched by calling the "sys/wrapping/unwrap"
+                          endpoint, passing in the wrappping token's ID. This
+                          is a numeric string with an optional suffix
+                          "s", "m", or "h"; if no suffix is specified it will
+                          be parsed as seconds. May also be specified via
                           VAULT_WRAP_TTL.
 `
 }